Modifications to make it work, and use a random TestRule

+++ cell.h
@@ -81,7 +81,7 @@ public:
     int i, j;
     Neighborhood neighborhood;
     BoundaryType boundary;
-    int state;
+    byte state;
 
 };
